Revert "Update Object.GetObject() to use JniRuntime.JniValueManager"
This reverts commit f3c122ee1b3d53918b990bbc9ac76b5fd4460499.

This doens't build, because d3cde470 turns all warnings into errors,
and this commit introduces a warning.

PR #9728 is the proper fix.

+++ src/Mono.Android/Java.Lang/Object.cs
@@ -156,9 +156,13 @@ namespace Java.Lang {
 			if (handle == IntPtr.Zero)
 				return null;
 
-			var p   = JNIEnvInit.ValueManager!.GetPeer (new JniObjectReference (handle), type);
-			JNIEnv.DeleteRef (handle, transfer);
-			return p;
+			var r = PeekObject (handle, type);
+			if (r != null) {
+				JNIEnv.DeleteRef (handle, transfer);
+				return r;
+			}
+
+			return Java.Interop.TypeManager.CreateInstance (handle, transfer, type);
 		}
 
 		[EditorBrowsable (EditorBrowsableState.Never)]
